requirements.txt
asgiref==3.11.1
Django==6.0.3
PyMySQL==1.1.2
sqlparse==0.5.5
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
django-filter==23.2